+++ lib/oxidized/node/stats.rb
@@ -1,7 +1,6 @@
 module Oxidized
   class Node
     class Stats
-      attr_accessor :mtime
       MAX_STAT = 10
 
       # @param [Job] job job whose information add to stats
@@ -36,15 +35,24 @@ module Oxidized
         @stats[:counter].reduce(0) { |m, h| h[0] == :success ? m : m + h[1] }
       end
 
+      def mtimes
+        @mtimes
+      end
+
+      def mtime
+        mtimes.last
+      end
+
       def update_mtime
-        @mtime = Time.now.utc
+        @mtimes.push Time.now.utc
+        @mtimes.shift if @mtimes.size > MAX_STAT
       end
 
       private
 
       def initialize
-        @mtime = "unknown"
-        @stats = {}
+        @mtimes = Array.new(MAX_STAT, "unknown")
+        @stats  = {}
         @stats[:counter] = Hash.new 0
       end
     end