+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 * FEATURE: add frr support to cumulus model (@User4574 / @bobthebutcher)
+* FEATURE: honour MAX_STAT in mtime, to store last N mtime
 
 ## 0.23.0